Attach verify_cert_ip_addresses() to IpAddrRef
