+++ src/end_entity.rs
@@ -18,7 +18,7 @@ use pki_types::{CertificateDer, SignatureVerificationAlgorithm, TrustAnchor, Uni
 
 use crate::crl::RevocationOptions;
 use crate::error::Error;
-use crate::subject_name::{verify_cert_dns_name, verify_cert_ip_addresses, SubjectNameRef};
+use crate::subject_name::{verify_cert_dns_name, SubjectNameRef};
 use crate::verify_cert::{self, KeyUsage};
 use crate::{cert, signed_data};
 
@@ -111,7 +111,7 @@ impl<'a> EndEntityCert<'a> {
     ) -> Result<(), Error> {
         match subject_name {
             SubjectNameRef::DnsName(dns_name) => verify_cert_dns_name(self, dns_name),
-            SubjectNameRef::IpAddress(ip_address) => verify_cert_ip_addresses(self, ip_address),
+            SubjectNameRef::IpAddress(ip_address) => ip_address.verify_cert_ip_addresses(self),
         }
     }
