+++ src/subject_name/mod.rs
@@ -29,7 +29,4 @@ pub use ip_address::{AddrParseError, IpAddrRef};
 pub use ip_address::IpAddr;
 
 mod verify;
-pub(super) use verify::{
-    check_name_constraints, verify_cert_dns_name, verify_cert_ip_addresses, GeneralName,
-    NameIterator,
-};
+pub(super) use verify::{check_name_constraints, verify_cert_dns_name, GeneralName, NameIterator};
